requirements.txt
fastapi==0.108.0
httpx==0.26.0
pytest==7.4.0
uvicorn==0.25.0
sqlmodel==0.0.16
passlib==1.7.4
python-jose==3.3.0
bcrypt==4.1.2
cryptography==42.0.2
python-multipart==0.0.9